Clippy: integration Vol. 2 (#439)
* clippy: a new portion of minor auto fixes
* clippy: initial configuration
* clippy: another auto fixes portion
* clippy: another auto fixes portion (with --broken-code)
- cargo clippy \
    --fix \
    --lib \
    -p kamu \
    --allow-dirty \
    --allow-staged \
    --broken-code
* clippy: manual build fixes
* clippy: another auto fixes portion for all rest crates (with --broken-code)
- cargo clippy \
    --fix \
    --all \
    --allow-dirty \
    --allow-staged \
    --broken-code
* Regenerate... generated files
* make lint: run clippy as well
* clippy: cover the rest warnings
* test_dataset_refs_conversions(): remove duplicated lines
* make lint-fix: introduce
* make codegen: do not specify extra edition (has read from config)
* Fixed a couple dozen Clippy warnings

---------

+++ src/app/cli/src/commands/list_command.rs
@@ -261,7 +261,7 @@ impl Command for ListCommand {
                 let mut blocks_stream = dataset.as_metadata_chain().iter_blocks();
                 while let Some((_, b)) = blocks_stream.try_next().await? {
                     if num_blocks == 0 {
-                        num_blocks = b.sequence_number as u64 + 1;
+                        num_blocks = b.sequence_number + 1;
                     }
                     if let Some(b) = b.as_data_stream_block() {
                         last_watermark = b.event.new_watermark.cloned();
